requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
django-rest-authtoken~=2.1.4
pillow~=10.4.0
drf-extra-fields~=3.7.0